Ajustados los métodos de creación de Eventos, Inscripciones y Usuarios para generar los IDs de los nuevos elementos como System.currentTimeStamp().

+++ app/src/main/java/com/ubung/tc/ubungmobile/modelo/persistencia/InterfazPersistencia.java
@@ -75,7 +75,7 @@ public interface InterfazPersistencia {
      * @throws ExcepcionPersistencia en caso que se presente algún error al guardar el evento.
      * @return El id del evento recién creado
      */
-    public int crearEvento(Date fechaHora, Zona zona, Deporte deporte, Usuario organizador)
+    public long crearEvento(Date fechaHora, Zona zona, Deporte deporte, Usuario organizador)
             throws  ExcepcionPersistencia;
 
     /**
